LayoutTests/platform/wpe/imported/w3c/web-platform-tests/xhr/send-authentication-basic-setrequestheader-existing-session-expected.txt


FAIL XMLHttpRequest: send() - "Basic" authenticated request using setRequestHeader() when there is an existing session assert_equals: expected "f8947137-d721-4624-8bb5-1e3c2d8ac873\npass" but got "7972fecd-cba5-44a4-9968-acb5b8db190f\nopen-pass"
Note: this test will only work as expected once per browsing session. Restart browser to re-test.
